⚡️Improve TypeSafety
