+++ src/models/auth/authModel.ts
@@ -1,21 +1,5 @@
 import { Schema, model } from 'mongoose';
-
-export interface IUser {
-    name: string;
-    email: string;
-    password: string;
-    role: string;
-    signInDate: Date;
-    failedLoginAttempts: string[];
-    lockUntil: Date;
-    isVerified: boolean;
-    verificationToken: number;
-    loginHistory: {
-        loginDate: Date;
-        loginIP: string;
-    }[];
-    passwordResetToken: string;
-}
+import { IUser } from '../../interface/auth.interface';
 
 const registerSchema = new Schema<IUser>(
     {
